Added parameter and return types to generated class
If the parameter or return type is not 'mixed', it will now add
it as a native return/parameter type.

All getter methods now have a nullable return type, because
we cannot be sure that the value has been set (even with required
types) when using get/set method strategy.